pyproject.toml
[project]
name = "toshima-tv-outage-notifier"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "tweepy>=4.14.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "responses>=0.24.0",
    "ruff>=0.9.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

# Ruff configuration
[tool.ruff]
# Python 3.13 をターゲット
target-version = "py313"

# 行の長さ制限（PEP 8 準拠）
line-length = 88

[tool.ruff.lint]
# 有効化するルールセット
select = [
    "E",     # pycodestyle errors
    "W",     # pycodestyle warnings
    "F",     # pyflakes
    "I",     # isort (import sorting)
    "N",     # pep8-naming
    "UP",    # pyupgrade
    "B",     # flake8-bugbear
    "C4",    # flake8-comprehensions
    "SIM",   # flake8-simplify
    "PIE",   # flake8-pie
    "RET",   # flake8-return
    "ARG",   # flake8-unused-arguments
]

# 無視するルール（プロジェクト特性に応じて調整）
ignore = [
    "E501",   # 行長制限（Black/Ruffフォーマッターに任せる）
    "SIM102", # ネストされたif文を許可（外側と内側で異なる意味を持つ場合がある）
]

# ルールごとの設定
[tool.ruff.lint.per-file-ignores]
# テストファイルでは一部ルールを緩和
"tests/**/*.py" = [
    "ARG001",  # Unused function argument (pytest fixtures)
    "ARG002",  # Unused method argument
]

[tool.ruff.lint.isort]
# import の並び順設定
known-first-party = ["src"]
section-order = ["future", "standard-library", "third-party", "first-party", "local-folder"]

[tool.ruff.format]
# フォーマッター設定
quote-style = "double"
indent-style = "space"
line-ending = "auto"
